# Conflicts:
#   server/src/main/java/server/QuizzzServer.java
#   server/src/main/java/server/controllers/ActionController.java
#   server/src/main/java/server/services/QuestionGenerator/ComparisonQuestionGenerator.java
#   server/src/main/java/server/services/QuestionGenerator/KnowledgeQuestionGenerator.java
#   server/src/main/java/server/services/QuestionGenerator/OpenQuestionGenerator.java
#   server/src/main/java/server/services/QuestionGenerator/QuestionGenerator.java

+++ server/src/main/java/server/services/QuestionGenerator/ComparisonQuestionGenerator.java
@@ -1,13 +1,13 @@
 package server.services.QuestionGenerator;
 
+import commons.Actions.Action;
+import commons.Actions.ActionCatalog;
 import commons.Questions.ComparisonQuestion;
-
 import commons.Questions.Question;
 import commons.Actions.*;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.ArrayList;
-
 import java.util.Collections;
 import java.util.List;
 import java.util.Random;
@@ -43,7 +43,7 @@ public class ComparisonQuestionGenerator {
 
         ComparisonQuestion comparisonQuestion = new ComparisonQuestion(makeComparisonStatement(sign), options);
 
-        return Pair.of(comparisonQuestion, Integer.toString(sign * Math.max(sign * firstAction.getConsumption(), Math.max(sign * secondAction.getConsumption(), sign * thirdAction.getConsumption()))));
+        return Pair.of(comparisonQuestion, Long.toString(sign * Math.max(sign * firstAction.getConsumption(), Math.max(sign * secondAction.getConsumption(), sign * thirdAction.getConsumption()))));
     }
 
     public static String makeComparisonStatement(int sign) {
